runelite-client: standardize plugin names
